bot/requirements.txt
absl-py==2.1.0
aioconsole==0.7.1
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.20.0
alluka==0.2.0
astunparse==1.6.3
attrs==23.2.0
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
cryptography==42.0.7
Flask==3.0.3
flatbuffers==24.3.25
frozenlist==1.4.1
gast==0.5.4
google-pasta==0.2.0
grpcio==1.63.0
h11==0.14.0
h2==4.1.0
h5py==3.11.0
hikari==2.0.0.dev125
hikari-lightbulb==2.3.5
hikari-miru==4.1.1
hpack==4.0.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.3.3
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
multidict==6.0.5
namex==0.0.8
nltk==3.8.1
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
outcome==1.3.0.post0
packaging==24.0
pillow==10.3.0
priority==2.0.0
protobuf==4.25.3
pycparser==2.22
pycryptodome==3.20.0
Pygments==2.18.0
python-dotenv==1.0.1
pytz==2024.1
Quart==0.19.5
quart-flask-patch==0.3.0
regex==2024.4.28
requests==2.31.0
rich==13.7.1
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
tqdm==4.66.4
trio==0.25.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4